requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
mysql-connector-python
python-dotenv
pandas
matplotlib
email_validator==1.3.1
WTForms
flask_mail
alembic
flask_login
openpyxl
